requirement.txt
python-dotenv
flask
flask_cors
authlib
requests
pyjwt
flask_jwt_extended